youtube_video_retrieve_service/requirements.txt
cachetools==5.3.1
certifi==2022.6.15
charset-normalizer==3.2.0
distlib==0.3.5
filelock==3.7.1
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-pubsub==2.18.4
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
numpy==1.26.1
oauthlib==3.2.2
pipenv==2022.7.4
platformdirs==2.5.2
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
taopypy==0.0.8
tweepy==4.14.0
uritemplate==4.1.1
urllib3==2.0.5
virtualenv==20.15.1
virtualenv-clone==0.5.7
